找到电影详情页界面卡顿原因(加载基础资料的时候)

+++ app/src/main/java/com/cicinnus/cateye/module/movie/find_movie/fixedboard_movie/oversea_movie/bean/OverseaHotMovieBean.java
@@ -165,7 +165,7 @@ public class OverseaHotMovieBean {
             private String overseaTime;
             private int pn;
             private boolean preShow;
-            private int proScore;
+            private double proScore;
             private int proScoreNum;
             private long pubDate;
             private int pubShowNum;
@@ -354,11 +354,11 @@ public class OverseaHotMovieBean {
                 this.preShow = preShow;
             }
 
-            public int getProScore() {
+            public double getProScore() {
                 return proScore;
             }
 
-            public void setProScore(int proScore) {
+            public void setProScore(double proScore) {
                 this.proScore = proScore;
             }
